# 028. Swagger (API UI)/api/swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "API for SHOP Website",
    "description": "This is a REST API application made with Express and documented with Swagger"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api"
    },
    {
      "url": "https://localhost:3000/api"
    },
    {
      "url": "http://shop.didan.id.vn/api"
    },
    {
      "url": "https://shop.didan.id.vn/api"
    }
  ],
...
  "paths": {
...
    "/payment/vnpay_return": {
      "get": {
        "tags": [
          "Payment"
        ],
        "description": "Endpoint return from VNPAY payment website",
        "parameters": [
          {
            "name": "vnp_SecureHash",
            "in": "query",
            "description": "Secure Hash from VNPAY",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "vnp_ResponseCode",
            "in": "query",
            "description": "Response Code from VNPAY",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Payment Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Payment Success"
                    },
                    "title": {
                      "type": "string",
                      "example": "Payment Success"
                    },
                    "path": {
                      "type": "string",
                      "example": "/checkout"
                    },
                    "code": {
                      "type": "string",
                      "example": "00"
                    }
                  },
                  "required": [
                    "message",
                    "title",
                    "path",
                    "code"
                  ],
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Payment Success"
                    },
                    "title": {
                      "type": "string",
                      "example": "Payment Success"
                    },
                    "path": {
                      "type": "string",
                      "example": "/checkout"
                    },
                    "code": {
                      "type": "string",
                      "example": "00"
                    }
                  },
                  "required": [
                    "message",
                    "title",
                    "path",
                    "code"
                  ],
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "308": {
            "description": "Permanent Redirect"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Payment Failed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Payment Failed"
                    },
                    "title": {
                      "type": "string",
                      "example": "Payment Failed"
                    },
                    "path": {
                      "type": "string",
                      "example": "/checkout"
                    },
                    "code": {
                      "type": "string",
                      "example": "97"
                    }
                  },
                  "required": [
                    "message",
                    "title",
                    "path",
                    "code"
                  ],
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Payment Failed"
                    },
                    "title": {
                      "type": "string",
                      "example": "Payment Failed"
                    },
                    "path": {
                      "type": "string",
                      "example": "/checkout"
                    },
                    "code": {
                      "type": "string",
                      "example": "97"
                    }
                  },
                  "required": [
                    "message",
                    "title",
                    "path",
                    "code"
                  ],
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Server Error"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Server Error"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
...
}